Added background and Prototypes to Minimap and finished Ability drag and drop

+++ Assets/NPCHealthController.cs
@@ -35,6 +35,6 @@ public class NPCHealthController : MonoBehaviour
 			health -= damage;
 		}
 		healthSlider.value = (float) health / (float) maxHealth;
-		healthText.text = health.ToString();
+		healthText.text = ((int)health).ToString();
 	}
 }
